Make dart_segment_get_info static inline
